fixed the mysterious duplicate jobs error

+++ prf_daemon.pl
@@ -52,7 +52,7 @@ if ($config->{checks}) {
     Check_Blast();
 }
 ## Some Arguments should be checked before others...
-if (defined($config->{jobs})) {
+if (defined($config->{make_jobs})) {
     Make_Queue_Jobs();
     exit(0);
 }    
@@ -1113,36 +1113,39 @@ sub CLEANUP {
 sub Make_Queue_Jobs {
     ## copying out of the perl script make_jobs.pl
     use Template;
-    my $config = new PRFConfig(config_file => "$ENV{PRFDB_HOME}/prfdb.conf");
-    chdir($config->{base});
-    my $template_config = $config;
-    $template_config->{PRE_PROCESS} = undef;
+    my $template_config;
+    $template_config->{PRE_PROCESS} = 0;
     $template_config->{EVAL_PERL} = 0;
     $template_config->{INTERPOLATE} = 0;
     $template_config->{POST_CHOMP} = 0;
     $template_config->{ABSOLUTE} = 1;
+    foreach my $k (keys %{$config}) {
+	next unless ($k =~ /^pbs/);
+	$template_config->{$k} = $config->{$k};
+    }
+    $template_config->{base} = $config->{base};
     my $template = new Template($template_config);
-    my $base = $template_config->{base};
     my $input_file = "descr/job_template";
     my @arches = split(/ /, $config->{pbs_arches});
     foreach my $arch (@arches) {
 	system("mkdir -p jobs/$arch") unless (-d "jobs/$arch");
 	my $archchar = substr($arch,0,3);
-	foreach my $daemon ("01" .. $config->{num_daemons}) {
+	foreach my $daemon ("01" .. $template_config->{pbs_num_daemons}) {
 	    my $output_file = "jobs/$arch/$daemon";
 	    ## First delete the existing job file in case a change has been made to the
 	    ## configuration file
 	    unlink($output_file);
 	    my $name = $template_config->{pbs_partialname};
 	    my $pbs_fullname = "${name}_${archchar}_${daemon}";
-	    my $incdir = "${base}/usr/perl.${arch}/lib";
+	    my $incdir = "$config->{base}/usr/perl.${arch}/lib";
 	    my $vars = {
 		pbs_shell => $template_config->{pbs_shell},
 		pbs_memory => $template_config->{pbs_memory},
 		pbs_cpu => $template_config->{pbs_cpu},
 		pbs_arch => $arch,
 		pbs_name => $pbs_fullname,
-		pbs_cput => $template_config->{pbs_cput},
+		pbs_cputime => $template_config->{pbs_cputime},
+		pbs_nodes => 1,
 		perl => $template_config->{perl},
 		incdir => $incdir,
 		daemon_name => $template_config->{daemon_name},
